Migration of Dashboard Canvas objects to VAS
Move ButtonObject to VAS.Drawing

Fixed some naming issues

Now TimerButtonLongomatch has no sense since is the same as VAS

Only CardObject and ScoreObject remins in Longomatch since it has no
sense to migrate to VAS

Migration DashboardCanvas to VAS

The class name in Longomatch has been renamed to SportDashboardCanvas
